Refactor/#43

+++ src/main/kotlin/com/example/days/domain/dailycheck/service/DailyCheckServiceImpl.kt
@@ -4,6 +4,7 @@ import com.example.days.domain.dailycheck.dto.request.DailyCheckRequest
 import com.example.days.domain.dailycheck.dto.response.DailyCheckResponse
 import com.example.days.domain.dailycheck.repository.DailyCheckRepository
 import com.example.days.domain.resolution.repository.ResolutionRepository
+import com.example.days.global.infra.security.AuthenticationUtil.getAuthenticationUserId
 import org.springframework.data.domain.Sort
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
@@ -15,29 +16,48 @@ class DailyCheckServiceImpl(
     private val resolutionRepository: ResolutionRepository
 ):DailyCheckService {
     @Transactional
-    override fun createDailyCheck(resolutionId: Long, request: DailyCheckRequest): DailyCheckResponse {
+    override fun createDailyCheck(resolutionId: Long, userId: Long, request: DailyCheckRequest): DailyCheckResponse {
+        val resolution = resolutionRepository.findByIdOrNull(resolutionId) ?: TODO("예외처리")
+
+        if(userId == resolution.author.id){
+            dailyCheckRepository.save(DailyCheckRequest.of(request, resolution))
+        }
         return resolutionRepository.findByIdOrNull(resolutionId)
             ?.let{ dailyCheckRepository.save(DailyCheckRequest.of(request, it)) }
             ?.let { DailyCheckResponse.from(it) }
             ?: TODO("예외처리")
     }
 
-    override fun getDailyCheckByList(resolutionId: Long): List<DailyCheckResponse> {
-        return dailyCheckRepository.findAll(Sort.by(Sort.Direction.DESC, "createdAt"))
-            .map { DailyCheckResponse.from(it) }
+    override fun getDailyCheckByList(resolutionId: Long, userId: Long): List<DailyCheckResponse> {
+        val resolution = resolutionRepository.findByIdOrNull(resolutionId) ?: TODO("예외처리")
+        if(userId == resolution.author.id){
+            return dailyCheckRepository.findAll(Sort.by(Sort.Direction.DESC, "createdAt"))
+                .map { DailyCheckResponse.from(it) }
+        }
+        else TODO()
     }
 
     @Transactional
-    override fun updateDailyCheck(resolutionId: Long, dailyCheckId: Long, request: DailyCheckRequest)
+    override fun updateDailyCheck(resolutionId: Long, userId: Long, dailyCheckId: Long, request: DailyCheckRequest)
     : DailyCheckResponse {
-        val resolution = resolutionRepository.findByIdOrNull(resolutionId) ?: TODO()
+        val resolution = resolutionRepository.findByIdOrNull(resolutionId) ?: TODO("예외처리")
         val dailyCheck = dailyCheckRepository.findByIdOrNull(dailyCheckId) ?: TODO()
-        dailyCheck.updateDailyCheck(request.memo, resolution)
-        return DailyCheckResponse.from(dailyCheck)
+        if(userId == resolution.author.id){
+            dailyCheck.updateDailyCheck(request.memo, resolution)
+            return DailyCheckResponse.from(dailyCheck)
+        }
+        else TODO()
+
+
     }
 
     @Transactional
-    override fun deleteDailyCheck(dailyCheckId: Long) {
-        dailyCheckRepository.deleteById(dailyCheckId)
+    override fun deleteDailyCheck(resolutionId: Long, dailyCheckId: Long, userId: Long) {
+        val resolution = resolutionRepository.findByIdOrNull(resolutionId) ?: TODO("예외처리")
+
+        if(userId == resolution.author.id) {
+            dailyCheckRepository.deleteById(dailyCheckId)
+        }
+        else TODO()
     }
 }
