+++ src/main/kotlin/com/example/days/domain/resolution/service/ResolutionServiceImpl.kt
@@ -1,21 +1,28 @@
 package com.example.days.domain.resolution.service
 
+import com.example.days.domain.category.repository.CategoryRepository
 import com.example.days.domain.resolution.dto.request.ResolutionRequest
 import com.example.days.domain.resolution.dto.response.ResolutionResponse
-import com.example.days.domain.resolution.model.Resolution
 import com.example.days.domain.resolution.repository.ResolutionRepository
+import com.example.days.domain.user.repository.UserRepository
+import com.example.days.global.common.SortOrder
+import org.springframework.data.domain.Page
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
+
 @Service
 class ResolutionServiceImpl(
-    private val resolutionRepository: ResolutionRepository
+    private val resolutionRepository: ResolutionRepository,
+    private val userRepository: UserRepository,
+    private val categoryRepository: CategoryRepository
 ): ResolutionService {
     @Transactional
-    override fun createResolution(request: ResolutionRequest): ResolutionResponse {
-        // TODO : 사용자 식별 가능한 로직 추가 예졍
-        val resolution = resolutionRepository.save(ResolutionRequest.of(request))
+    override fun createResolution(request: ResolutionRequest, userId: Long): ResolutionResponse {
+        val user = userRepository.findByIdOrNull(userId) ?: TODO()
+        val category = categoryRepository.findByName(request.category) ?: TODO()
+        val resolution = resolutionRepository.save(ResolutionRequest.of(request, category, user))
         return ResolutionResponse.from(resolution)
     }
 
@@ -24,23 +31,30 @@ class ResolutionServiceImpl(
         return ResolutionResponse.from(resolution)
     }
 
-    override fun getResolutionList(): List<ResolutionResponse> {
-        return resolutionRepository.findAll().map{ResolutionResponse.from(it)}
+    override fun getResolutionListPaginated(page: Int, sortOrder: SortOrder?): Page<ResolutionResponse> {
+        val resolutionList = resolutionRepository.findByPageable(page, sortOrder)
+        return resolutionList.map { ResolutionResponse.from(it) }
     }
 
     @Transactional
-    override fun updateResolution(resolutionId: Long, request: ResolutionRequest): ResolutionResponse {
-        // TODO : 목표 작성자만 수정 가능하도록 제한
+    override fun updateResolution(resolutionId: Long, userId: Long, request: ResolutionRequest): ResolutionResponse {
+        val category = categoryRepository.findByName(request.category) ?: TODO()
         val updatedResolution = getByIdOrNull(resolutionId)
-        updatedResolution.updateResolution(request.title, request.description, request.category)
-        return ResolutionResponse.from(updatedResolution)
+        if(updatedResolution.author.id == userId){
+            updatedResolution.updateResolution(request.title, request.description, category)
+            return ResolutionResponse.from(updatedResolution)
+        }
+        else TODO("예외처리")
+
     }
 
     @Transactional
-    override fun deleteResolution(resolutionId: Long) {
-        // TODO : 목표 작성자만 삭제 가능하도록 제한
+    override fun deleteResolution(resolutionId: Long, userId: Long) {
         val resolution = getByIdOrNull(resolutionId)
-        resolutionRepository.delete(resolution)
+        if (resolution.author.id == userId){
+            resolutionRepository.delete(resolution)
+        }
+        else TODO("예외처리")
     }
 
     fun getByIdOrNull(id: Long) = resolutionRepository.findByIdOrNull(id) ?: TODO("예외처리 구현예정")
